fix : refund refundAmt 타입 변경 by 우영

+++ back/app/src/main/java/com/sloway/app/payment/refund/dto/response/RefundResDto.java
@@ -7,6 +7,7 @@ import com.sloway.app.payment.refund.entity.RefundEntity;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -17,7 +18,7 @@ public class RefundResDto {
     private Long payNo;
     private Long rsvnNo;
     private Long ucNo;
-    private Integer refundAmt;
+    private BigDecimal refundAmt;
     private RefundRate refundRate;
     private RefundReason refundReason;
     private LocalDateTime requestedAt;
